# packages/contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JB API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/discovery/search": {
   "post": {
    "tags": [
     "discovery"
    ],
    "summary": "Search",
    "operationId": "search_api_v1_discovery_search_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SearchIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchJobAccepted"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/discovery/jobs/{job_id}": {
   "get": {
    "tags": [
     "discovery"
    ],
    "summary": "Job Stream",
    "operationId": "job_stream_api_v1_discovery_jobs__job_id__get",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/{mid}": {
   "get": {
    "tags": [
     "markets"
    ],
    "summary": "Get Market",
    "operationId": "get_market_api_v1_markets__mid__get",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Mid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MarketSummary"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/{mid}/book": {
   "get": {
    "tags": [
     "markets"
    ],
    "summary": "Get Book",
    "operationId": "get_book_api_v1_markets__mid__book_get",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Mid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Book"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/{mid}/orders": {
   "post": {
    "tags": [
     "markets"
    ],
    "summary": "Place Order",
    "operationId": "place_order_api_v1_markets__mid__orders_post",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Mid"
      }
     },
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OrderIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Order"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/{mid}/orders/mine": {
   "get": {
    "tags": [
     "markets"
    ],
    "summary": "My Orders",
    "operationId": "my_orders_api_v1_markets__mid__orders_mine_get",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Mid"
      }
     },
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Order"
         },
         "title": "Response My Orders Api V1 Markets  Mid  Orders Mine Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/orders/{oid}": {
   "delete": {
    "tags": [
     "markets"
    ],
    "summary": "Cancel Order",
    "operationId": "cancel_order_api_v1_markets_orders__oid__delete",
    "parameters": [
     {
      "name": "oid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Oid"
      }
     },
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Order"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/{mid}/batches": {
   "get": {
    "tags": [
     "markets"
    ],
    "summary": "Batches",
    "description": "Price history. By default only rounds that carry a price (trades or limit up/down steps),\nwhich is what the chart wants; `all=true` includes quiet rounds with clearing_price null.",
    "operationId": "batches_api_v1_markets__mid__batches_get",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Mid"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 60,
       "title": "Limit"
      }
     },
     {
      "name": "snapshot",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Snapshot"
      }
     },
     {
      "name": "all",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "All"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Batch"
         },
         "title": "Response Batches Api V1 Markets  Mid  Batches Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/{mid}/trades": {
   "get": {
    "tags": [
     "markets"
    ],
    "summary": "Trades",
    "operationId": "trades_api_v1_markets__mid__trades_get",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Mid"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Trade"
         },
         "title": "Response Trades Api V1 Markets  Mid  Trades Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/{mid}/batch/run": {
   "post": {
    "tags": [
     "markets"
    ],
    "summary": "Run Batch Now",
    "description": "Force a round to clear now. Demo and test convenience.",
    "operationId": "run_batch_now_api_v1_markets__mid__batch_run_post",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Mid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Batch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/markets/{mid}/narrative": {
   "get": {
    "tags": [
     "markets"
    ],
    "summary": "Narrative",
    "description": "Two lines of Grok tape commentary for this market, refreshed when a new round clears.",
    "operationId": "narrative_api_v1_markets__mid__narrative_get",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Mid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Band": {
    "properties": {
     "pct": {
      "type": "number",
      "title": "Pct"
     },
     "low": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Low"
     },
     "high": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "High"
     }
    },
    "type": "object",
    "required": [
     "pct",
     "low",
     "high"
    ],
    "title": "Band"
   },
   "Batch": {
    "properties": {
     "_id": {
      "type": "string",
      "title": "Id"
     },
     "market_id": {
      "type": "string",
      "title": "Market Id"
     },
     "t": {
      "type": "string",
      "title": "T"
     },
     "clearing_price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Clearing Price"
     },
     "volume": {
      "type": "number",
      "title": "Volume"
     },
     "imbalance": {
      "type": "number",
      "title": "Imbalance"
     },
     "n_buy": {
      "type": "integer",
      "title": "N Buy"
     },
     "n_sell": {
      "type": "integer",
      "title": "N Sell"
     },
     "band_hit": {
      "type": "boolean",
      "title": "Band Hit"
     },
     "ref_moved": {
      "type": "boolean",
      "title": "Ref Moved"
     },
     "book_snapshot": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Book Snapshot"
     }
    },
    "type": "object",
    "required": [
     "_id",
     "market_id",
     "t",
     "clearing_price",
     "volume",
     "imbalance",
     "n_buy",
     "n_sell",
     "band_hit",
     "ref_moved"
    ],
    "title": "Batch"
   },
   "Belief": {
    "properties": {
     "mu": {
      "type": "number",
      "title": "Mu"
     },
     "sigma": {
      "type": "number",
      "title": "Sigma"
     },
     "s_m": {
      "type": "number",
      "title": "S M"
     },
     "n_rounds": {
      "type": "integer",
      "title": "N Rounds"
     },
     "model_value": {
      "type": "number",
      "title": "Model Value"
     },
     "market_value": {
      "type": "number",
      "title": "Market Value"
     }
    },
    "type": "object",
    "required": [
     "mu",
     "sigma",
     "s_m",
     "n_rounds",
     "model_value",
     "market_value"
    ],
    "title": "Belief"
   },
   "Book": {
    "properties": {
     "market_id": {
      "type": "string",
      "title": "Market Id"
     },
     "bids": {
      "items": {
       "$ref": "#/components/schemas/BookLevel"
      },
      "type": "array",
      "title": "Bids"
     },
     "asks": {
      "items": {
       "$ref": "#/components/schemas/BookLevel"
      },
      "type": "array",
      "title": "Asks"
     },
     "last": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last"
     },
     "ref": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref"
     },
     "indicative_price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Indicative Price"
     },
     "next_batch_at": {
      "type": "string",
      "title": "Next Batch At"
     },
     "band": {
      "$ref": "#/components/schemas/Band"
     },
     "halted": {
      "type": "boolean",
      "title": "Halted"
     },
     "n_open_orders": {
      "type": "integer",
      "title": "N Open Orders"
     }
    },
    "type": "object",
    "required": [
     "market_id",
     "bids",
     "asks",
     "last",
     "ref",
     "indicative_price",
     "next_batch_at",
     "band",
     "halted",
     "n_open_orders"
    ],
    "title": "Book"
   },
   "BookLevel": {
    "properties": {
     "price": {
      "type": "number",
      "title": "Price"
     },
     "qty": {
      "type": "number",
      "title": "Qty"
     },
     "origin": {
      "type": "string",
      "enum": [
       "user",
       "treasury",
       "bot",
       "agent"
      ],
      "title": "Origin"
     }
    },
    "type": "object",
    "required": [
     "price",
     "qty",
     "origin"
    ],
    "title": "BookLevel"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "Level": {
    "properties": {
     "price": {
      "type": "number",
      "title": "Price"
     },
     "qty": {
      "type": "number",
      "title": "Qty"
     }
    },
    "type": "object",
    "required": [
     "price",
     "qty"
    ],
    "title": "Level"
   },
   "MarketSummary": {
    "properties": {
     "shares_outstanding": {
      "type": "integer",
      "title": "Shares Outstanding"
     },
     "float": {
      "type": "number",
      "title": "Float"
     },
     "retained": {
      "type": "number",
      "title": "Retained"
     },
     "tick": {
      "type": "number",
      "title": "Tick"
     },
     "last_price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Price"
     },
     "ref_price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref Price"
     },
     "batch_interval_s": {
      "type": "number",
      "title": "Batch Interval S"
     },
     "next_batch_at": {
      "type": "string",
      "title": "Next Batch At"
     },
     "band_pct": {
      "type": "number",
      "title": "Band Pct"
     },
     "belief": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Belief"
       },
       {
        "type": "null"
       }
      ]
     },
     "treasury": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Treasury"
       },
       {
        "type": "null"
       }
      ]
     },
     "fees_collected": {
      "type": "number",
      "title": "Fees Collected"
     },
     "halted": {
      "type": "boolean",
      "title": "Halted"
     }
    },
    "type": "object",
    "required": [
     "shares_outstanding",
     "float",
     "retained",
     "tick",
     "last_price",
     "ref_price",
     "batch_interval_s",
     "next_batch_at",
     "band_pct",
     "belief",
     "treasury",
     "fees_collected",
     "halted"
    ],
    "title": "MarketSummary"
   },
   "Order": {
    "properties": {
     "_id": {
      "type": "string",
      "title": "Id"
     },
     "market_id": {
      "type": "string",
      "title": "Market Id"
     },
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "side": {
      "type": "string",
      "enum": [
       "buy",
       "sell"
      ],
      "title": "Side"
     },
     "qty": {
      "type": "number",
      "title": "Qty"
     },
     "limit_price": {
      "type": "number",
      "title": "Limit Price"
     },
     "status": {
      "type": "string",
      "enum": [
       "open",
       "filled",
       "partial",
       "cancelled"
      ],
      "title": "Status"
     },
     "filled_qty": {
      "type": "number",
      "title": "Filled Qty"
     },
     "origin": {
      "type": "string",
      "enum": [
       "user",
       "treasury",
       "bot",
       "agent"
      ],
      "title": "Origin"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     },
     "cancelled_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cancelled At"
     }
    },
    "type": "object",
    "required": [
     "_id",
     "market_id",
     "user_id",
     "side",
     "qty",
     "limit_price",
     "status",
     "filled_qty",
     "origin",
     "created_at",
     "cancelled_at"
    ],
    "title": "Order"
   },
   "OrderIn": {
    "properties": {
     "side": {
      "type": "string",
      "enum": [
       "buy",
       "sell"
      ],
      "title": "Side"
     },
     "qty": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Qty"
     },
     "limit_price": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Limit Price"
     }
    },
    "type": "object",
    "required": [
     "side",
     "qty",
     "limit_price"
    ],
    "title": "OrderIn"
   },
...
   "SearchIn": {
    "properties": {
     "q": {
      "type": "string",
      "title": "Q"
     }
    },
    "type": "object",
    "required": [
     "q"
    ],
    "title": "SearchIn"
   },
   "SearchIntent": {
    "properties": {
     "category": {
      "type": "string",
      "title": "Category"
     },
     "naics_guess": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Naics Guess"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "min_value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Value"
     },
     "max_value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Value"
     },
     "must_have": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Must Have"
     }
    },
    "type": "object",
    "required": [
     "category",
     "naics_guess",
     "state",
     "city",
     "min_value",
     "max_value",
     "must_have"
    ],
    "title": "SearchIntent"
   },
   "SearchJobAccepted": {
    "properties": {
     "job_id": {
      "type": "string",
      "title": "Job Id"
     },
     "intent": {
      "$ref": "#/components/schemas/SearchIntent"
     }
    },
    "type": "object",
    "required": [
     "job_id",
     "intent"
    ],
    "title": "SearchJobAccepted"
   },
...
   "Trade": {
    "properties": {
     "_id": {
      "type": "string",
      "title": "Id"
     },
     "market_id": {
      "type": "string",
      "title": "Market Id"
     },
     "batch_id": {
      "type": "string",
      "title": "Batch Id"
     },
     "buyer_id": {
      "type": "string",
      "title": "Buyer Id"
     },
     "seller_id": {
      "type": "string",
      "title": "Seller Id"
     },
     "qty": {
      "type": "number",
      "title": "Qty"
     },
     "price": {
      "type": "number",
      "title": "Price"
     },
     "t": {
      "type": "string",
      "title": "T"
     }
    },
    "type": "object",
    "required": [
     "_id",
     "market_id",
     "batch_id",
     "buyer_id",
     "seller_id",
     "qty",
     "price",
     "t"
    ],
    "title": "Trade"
   },
   "Treasury": {
    "properties": {
     "unsold_float": {
      "type": "number",
      "title": "Unsold Float"
     },
     "proceeds": {
      "type": "number",
      "title": "Proceeds"
     },
     "floor_price": {
      "type": "number",
      "title": "Floor Price"
     },
     "floor_qty": {
      "type": "number",
      "title": "Floor Qty"
     },
     "bought_back": {
      "type": "number",
      "title": "Bought Back"
     },
     "ask_ladder": {
      "items": {
       "$ref": "#/components/schemas/Level"
      },
      "type": "array",
      "title": "Ask Ladder"
     }
    },
    "type": "object",
    "required": [
     "unsold_float",
     "proceeds",
     "floor_price",
     "floor_qty",
     "bought_back",
     "ask_ladder"
    ],
    "title": "Treasury"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}